+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #893 Allow Client contact to access to batches
 
 **Changed**